Changing ATTR_PG from 32 to 16 bit
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29834
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Dev-Ready: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29836
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/hwpf/src/plat/target.C
@@ -293,7 +293,7 @@ namespace fapi2
                                           bool & o_present)
      {
         o_present = false;
-        uint32_t attr_value = 0;
+        uint16_t attr_value = 0;
         FAPI_ATTR_GET(fapi2::ATTR_PG,
                 i_target,
                 attr_value);
@@ -472,7 +472,7 @@ fapi_try_exit:
             fapi2::Target<fapi2::TARGET_TYPE_EQ> l_parent = target_name.getParent<fapi2::TARGET_TYPE_EQ>();
 
             // Get the parent EQ's ATTR_PG
-            uint32_t l_eqAttrPg = 0;
+            uint16_t l_eqAttrPg = 0;
             FAPI_ATTR_GET(fapi2::ATTR_PG, l_parent.getParent<TARGET_TYPE_PERV>(), l_eqAttrPg);
 
             // Check if this EX's L2 and L3 regions are marked "good"
@@ -507,7 +507,7 @@ fapi_try_exit:
             fapi2::Target<fapi2::TARGET_TYPE_PERV>
                 l_nestTarget((plat_getTargetHandleByChipletNumber<TARGET_TYPE_PERV>(N3_CHIPLET - (MCS_PER_MCBIST * (i / MCS_PER_MCBIST)))));
 
-            uint32_t l_attrPg = 0;
+            uint16_t l_attrPg = 0;
 
             FAPI_ATTR_GET(fapi2::ATTR_PG, l_nestTarget, l_attrPg);
